move get_loss_factories
