+++ lightning_pose/api/model.py
@@ -109,13 +109,13 @@ def load_model_from_checkpoint(
     """
     if ckpt_file is None:
         raise ValueError('ckpt_file must be provided to load a model from checkpoint')
+    from lightning_pose.losses import get_loss_factories
     from lightning_pose.models import check_if_semi_supervised
     from lightning_pose.utils.io import return_absolute_data_paths
     from lightning_pose.utils.scripts import (
         get_data_module,
         get_dataset,
         get_imgaug_transform,
-        get_loss_factories,
     )
 
     delete_extras = False
