+++ lightning_pose/losses/__init__.py
@@ -1,1 +1,3 @@
 """Supervised and semi-supervised loss functions for pose estimation."""
+
+from lightning_pose.losses.factory import get_loss_classes, get_loss_factories
