+++ lightning_pose/utils/scripts.py
@@ -41,7 +41,6 @@ __all__ = [
     "get_imgaug_transform",
     "get_dataset",
     "get_data_module",
-    "get_loss_factories",
     "get_model",
     "get_callbacks",
     "calculate_steps_per_epoch",
@@ -256,139 +255,6 @@ def get_data_module(
             imgaug=cfg.training.get("imgaug", "default"),
         )
     return data_module
-
-
-def get_loss_factories(
-    cfg: DictConfig | ListConfig,
-    data_module: BaseDataModule | UnlabeledDataModule,
-) -> dict:
-    """Create loss factory that orchestrates different losses during training."""
-
-    cfg_loss_dict = OmegaConf.to_object(cfg.losses)
-    assert cfg_loss_dict is not None
-
-    loss_params_dict = {"supervised": {}, "unsupervised": {}}
-
-    # collect all supervised losses in a dict; no extra params needed
-    # set "log_weight = 0.0" so that weight = 1 and effective weight is (1 / 2)
-    if cfg.model.model_type.find("heatmap") > -1:
-        loss_name = "heatmap_" + cfg.model.heatmap_loss_type
-        loss_params_dict["supervised"][loss_name] = {"log_weight": 0.0}
-        if cfg.model.model_type.find("multiview") > -1 and cfg.data.get("camera_params_file"):
-
-            log_weight_sp = cfg.losses.get(
-                "supervised_pairwise_projections", {}
-            ).get("log_weight")
-            if log_weight_sp is not None:
-                print("Adding supervised pairwise projection loss")
-                loss_params_dict["supervised"]["supervised_pairwise_projections"] = {
-                    "log_weight": log_weight_sp
-                }
-
-            log_weight_hr = cfg.losses.get(
-                "supervised_reprojection_heatmap_mse", {}
-            ).get("log_weight")
-            if log_weight_hr is not None:
-                print("Adding supervised reprojection heatmap loss")
-                height_og = cfg.data.image_resize_dims.height
-                width_og = cfg.data.image_resize_dims.width
-                height_ds = int(height_og // (2 ** cfg.data.get("downsample_factor", 2)))
-                width_ds = int(width_og // (2 ** cfg.data.get("downsample_factor", 2)))
-                loss_params_dict["supervised"]["supervised_reprojection_heatmap_mse"] = {
-                    "log_weight": log_weight_hr,
-                    "original_image_height": height_og,
-                    "original_image_width": width_og,
-                    "downsampled_image_height": height_ds,
-                    "downsampled_image_width": width_ds,
-                }
-
-    else:
-        loss_params_dict["supervised"][cfg.model.model_type] = {"log_weight": 0.0}
-
-    # collect all unsupervised losses and their params in a dict
-    if cfg.model.losses_to_use is not None:
-        for loss_name in cfg.model.losses_to_use:
-            # general parameters
-            loss_params_dict["unsupervised"][loss_name] = cfg_loss_dict[loss_name]
-            loss_params_dict["unsupervised"][loss_name]["loss_name"] = loss_name
-            # loss-specific parameters
-            if loss_name[:8] == "unimodal" or loss_name[:16] == "temporal_heatmap":
-                if cfg.model.model_type == "regression":
-                    raise NotImplementedError(
-                        "unimodal loss can only be used with classes inheriting from "
-                        "HeatmapTracker. \nYou specified a RegressionTracker model."
-                    )
-                # record original image dims (after initial resizing)
-                raise Exception(
-                    "need to update unimodal and/or temporal heatmap loss to not use "
-                    "cfg.data.image_resize_dims, which has been deprecated."
-                )
-                height_og = cfg.data.image_resize_dims.height
-                width_og = cfg.data.image_resize_dims.width
-                loss_params_dict["unsupervised"][loss_name][
-                    "original_image_height"
-                ] = height_og
-                loss_params_dict["unsupervised"][loss_name][
-                    "original_image_width"
-                ] = width_og
-                # record downsampled image dims
-                height_ds = int(height_og // (2 ** cfg.data.get("downsample_factor", 2)))
-                width_ds = int(width_og // (2 ** cfg.data.get("downsample_factor", 2)))
-                loss_params_dict["unsupervised"][loss_name][
-                    "downsampled_image_height"
-                ] = height_ds
-                loss_params_dict["unsupervised"][loss_name][
-                    "downsampled_image_width"
-                ] = width_ds
-                if loss_name[:8] == "unimodal":
-                    loss_params_dict["unsupervised"][loss_name][
-                        "uniform_heatmaps"
-                    ] = cfg.training.get("uniform_heatmaps_for_nan_keypoints", False)
-            elif loss_name == "pca_multiview":
-                if cfg.data.get("view_names", None) and len(cfg.data.view_names) > 1:
-                    # assume user has provided a set of columns that are present in each view
-                    num_keypoints = cfg.data.num_keypoints
-                    num_views = len(cfg.data.view_names)
-                    if isinstance(cfg.data.mirrored_column_matches[0], int):
-                        loss_params_dict["unsupervised"][loss_name][
-                            "mirrored_column_matches"
-                        ] = [
-                            (v * num_keypoints
-                             + np.array(cfg.data.mirrored_column_matches, dtype=int)).tolist()
-                            for v in range(num_views)
-                        ]
-                    else:
-                        # allow user to force specific mapping in multiview case
-                        loss_params_dict["unsupervised"][loss_name][
-                            "mirrored_column_matches"
-                        ] = cfg.data.mirrored_column_matches
-                else:
-                    # user must provide all matching columns
-                    loss_params_dict["unsupervised"][loss_name][
-                        "mirrored_column_matches"
-                    ] = cfg.data.mirrored_column_matches
-            elif loss_name == "pca_singleview":
-                if cfg.data.get("view_names", None) and len(cfg.data.view_names) > 1:
-                    raise NotImplementedError(
-                        "The Pose PCA loss is currently not implemented for multiview data."
-                    )
-                else:
-                    loss_params_dict["unsupervised"][loss_name][
-                        "columns_for_singleview_pca"
-                    ] = cfg.data.get('columns_for_singleview_pca', None)
-
-    # build supervised loss factory, which orchestrates all supervised losses
-    loss_factory_sup = LossFactory(
-        losses_params_dict=loss_params_dict["supervised"],
-        data_module=data_module,
-    )
-    # build unsupervised loss factory, which orchestrates all unsupervised losses
-    loss_factory_unsup = LossFactory(
-        losses_params_dict=loss_params_dict["unsupervised"],
-        data_module=data_module,
-    )
-
-    return {"supervised": loss_factory_sup, "unsupervised": loss_factory_unsup}
 
 
 def get_model(
