+++ tests/utils/test_predictions.py
@@ -7,8 +7,9 @@ import pandas as pd
 import pytest
 from omegaconf import OmegaConf
 
+from lightning_pose.losses import get_loss_factories
 from lightning_pose.utils.predictions import make_dlc_pandas_index, predict_dataset
-from lightning_pose.utils.scripts import get_loss_factories, get_model
+from lightning_pose.utils.scripts import get_model
 
 
 class TestMakeDlcPandasIndex:
